Services cleanup, rename and repackage.
