+++ app/src/main/java/pl/swislowski/kamil/project/platerecognition/android/service/RecognitionRegistrationPlateService.java
@@ -1,6 +1,5 @@
 package pl.swislowski.kamil.project.platerecognition.android.service;
 
-import android.annotation.SuppressLint;
 import android.util.Log;
 
 import java.io.ByteArrayOutputStream;
@@ -10,21 +9,19 @@ import java.io.InputStream;
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
 import okhttp3.RequestBody;
-import pl.swislowski.kamil.project.platerecognition.spring.web.model.RegistrationPlateModel;
 import retrofit2.Call;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
-public class RecognitionRegistrationPlateService {
+import pl.swislowski.kamil.project.platerecognition.spring.web.model.RegistrationPlateModel;
 
-    private static final String TAG = "RecognitionRegistrationPlateService";
+public class RecognitionRegistrationPlateService {
+    private static final String TAG = "RRPService";
     public static final String SERVER_URL = "http://10.0.2.2:8080/";
     public static final String HEROKU_SERVER_URL = "https://plate-recognition-spring.herokuapp.com/";
 
-    @SuppressLint("LongLogTag")
-    public static RegistrationPlateModel recognition(InputStream inputStream, String fileName) throws IOException {
-
-        Log.i(TAG, "########Starting recognition method ...");
+    public static RegistrationPlateModel recognitionAsync(InputStream inputStream) throws IOException {
+        Log.i(TAG, "#### Starting recognitionAsync method ...");
 
         byte[] buffor = new byte[512];
         ByteArrayOutputStream upload = new ByteArrayOutputStream();
@@ -34,7 +31,7 @@ public class RecognitionRegistrationPlateService {
 
         MultipartBody.Part filePart = MultipartBody.Part.createFormData(
                 "upload",
-                fileName,
+                "file.jpg",
                 RequestBody.create(MediaType.parse("image/*"), upload.toByteArray()));
 
         Retrofit retrofit = new Retrofit.Builder()
@@ -42,11 +39,13 @@ public class RecognitionRegistrationPlateService {
                 .addConverterFactory(GsonConverterFactory.create())
                 .build();
 
-        RecognitionRegistrationPlateRetrofitService recognitionRegistrationPlateRetrofitService =
-                retrofit.create(RecognitionRegistrationPlateRetrofitService.class);
+        RecognitionRegistrationPlateRetrofitService retrofitService = retrofit.create(RecognitionRegistrationPlateRetrofitService.class);
+        Call<RegistrationPlateModel> call = retrofitService.recognizeAsync(filePart);
 
-        Call<RegistrationPlateModel> call = recognitionRegistrationPlateRetrofitService.recognize(filePart);
+        RegistrationPlateModel body = call.execute().body();
+//        call.enqueue(new RRPRetrofitCallback());
+        Log.i(TAG, "#### recognitionAsync body: " + body);
 
-        return call.execute().body();
+        return body;
     }
 }
